Allow excluding sub pull requests from the description
Useful for projects with complex workflows, with a branching model with
multiple levels. Example:

* master
|
|
 \*staging
 |
 |
 |\*feature1
 | |
 | |\*sub-feature-1
 | |
 | |
 | |\*sub-feature-2
 |\*feature2

In this example, each branch is merged to its "parent" with a pull request.
By default, if we want to release staging to master, it will include the
pull requests for feature1 and feature2, but it will also include the
pull requests to sub-feature1 and sub-feature2, as they are part of
feature1. With this new option, it is possible to tell the releaser to
only include the pull requests sent directly to the head branch. That is,
only feature1 and feature 2